fix repeat data issue & adjust file structure
