+++ public/index.html
@@ -10,11 +10,11 @@
     <meta name="apple-mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-status-bar-style" content="black">
     <meta name="apple-mobile-web-app-title" content="Weather PWA">
-    <link rel="apple-touch-icon" href="images/icons/icon-152x152.png">
-    <meta name="msapplication-TileImage" content="images/icons/icon-144x144.png">
+    <link rel="apple-touch-icon" href="<%= BASE_URL %>assets/icons/icon-152x152.png">
+    <meta name="msapplication-TileImage" content="<%= BASE_URL %>assets/icons/icon-144x144.png">
     <meta name="msapplication-TileColor" content="#2f3ba2">
     <title>Weather PWA</title>
-    <link rel="manifest" href="./manifest.json">
+    <link rel="manifest" href="<%= BASE_URL %>manifest.json">
   </head>
   <body>
     <noscript>
