+++ public/manifest.json
@@ -1,24 +1,24 @@
 {
-  "name": "Weather",
+  "name": "Weather PWA",
   "short_name": "Weather",
   "icons": [{
-    "src": "images/icons/icon-128x128.png",
+    "src": "assets/icons/icon-128x128.png",
       "sizes": "128x128",
       "type": "image/png"
     }, {
-      "src": "images/icons/icon-144x144.png",
+      "src": "assets/icons/icon-144x144.png",
       "sizes": "144x144",
       "type": "image/png"
     }, {
-      "src": "images/icons/icon-152x152.png",
+      "src": "assets/icons/icon-152x152.png",
       "sizes": "152x152",
       "type": "image/png"
     }, {
-      "src": "images/icons/icon-192x192.png",
+      "src": "assets/icons/icon-192x192.png",
       "sizes": "192x192",
       "type": "image/png"
     }, {
-      "src": "images/icons/icon-256x256.png",
+      "src": "assets/icons/icon-256x256.png",
       "sizes": "256x256",
       "type": "image/png"
     }],
