+++ public/service-worker.js
@@ -1,0 +1,69 @@
+var dataCacheName = 'weatherData-v1';
+var cacheName = 'weatherPWA';
+var filesToCache = [
+  '/',
+  '/index.html',
+  '/scripts/app.js',
+  '/styles/inline.css',
+  '/images/clear.png',
+  '/images/cloudy-scattered-showers.png',
+  '/images/cloudy.png',
+  '/images/fog.png',
+  '/images/ic_add_white_24px.svg',
+  '/images/ic_refresh_white_24px.svg',
+  '/images/partly-cloudy.png',
+  '/images/rain.png',
+  '/images/scattered-showers.png',
+  '/images/sleet.png',
+  '/images/snow.png',
+  '/images/thunderstorm.png',
+  '/images/wind.png'
+];
+
+self.addEventListener('install', function(e) {
+  console.log('[serviceWorker] install');
+  e.waitUntil(
+    caches.open(cacheName).then(function(cache) {
+      console.log('[serviceWorker] Caching app shell');
+      return cache.addAll(filesToCache);
+    })
+  );
+});
+
+self.addEventListener('activate', function(e) {
+  console.log('[serviceWorker] Activate');
+  e.waitUntil(
+    caches.keys().then(function(keyList) {
+      return Promise.all(keyList.map(function(key) {
+        console.log('[serviceWorker] removing old cache', key);
+        if (key !== cacheName && key !== dataCacheName) {
+          return caches.delete(key);
+        }
+      }));
+    })
+  );
+});
+
+self.addEventListener('fetch', function(e) {
+  console.log('[serviceWorker] Fetch', e.request.url);
+  var dataUrl = 'http://tomo.club/';
+  if (e.request.url.indexOf(dataUrl) > -1) {
+    e.respondWith(
+      fetch(e.request)
+        .then(function(response) {
+          return caches.open(dataCacheName).then(function(cache) {
+            cache.put(e.request.url, response.clone());
+            console.log('[serviceWorker] Fetched & Cached Data');
+            return response;
+          });
+        })
+    );
+  } else {
+    e.respondWith(
+      caches.match(e.request).then(function(response) {
+        console.log('[serviceWorker] network Fetch');
+        return response || fetch(e.request);
+      })
+    );
+  }
+});
